Remove dead BoxModel and Rules state from Redux
