--- src/devtools/client/inspector/rules/selectors/rules.ts
@@ -1,3 +1,0 @@
-import { UIState } from "ui/state";
-
-export const getRules = (state: UIState) => state.rules.rules;
